Fix CelebA datasets, remove eval() call
