+++ utils/data/__init__.py
@@ -2,3 +2,4 @@ from .dataloader import *
 from .gaussian_process import *
 from .helpers import DIR_DATA
 from .imgs import *
+from .datasets import *
